# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ServiceResponse_null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"anyOf": [],
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"AppointmentResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"purpose": {
						"type": "string"
					},
					"host": {
						"type": "string"
					},
					"department": {
						"type": "string"
					},
					"appointmentDate": {
						"type": "string"
					},
					"appointmentTime": {
						"type": "string"
					},
					"timeDuration": {
						"type": "string"
					},
					"appointmentLocation": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ServiceResponse_AppointmentResponse-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"items": {
							"$ref": "#/components/schemas/AppointmentResponse"
						},
						"type": "array",
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ServiceResponse_AppointmentResponse-or-null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"allOf": [
							{
								"$ref": "#/components/schemas/AppointmentResponse"
							}
						],
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AppointmentCreateAttributes": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"purpose": {
						"type": "string"
					},
					"host": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"pending",
							"confirmed",
							"canceled",
							"onhold",
							"completed"
						]
					},
					"department": {
						"type": "string"
					},
					"company": {
						"type": "string"
					},
					"appointmentDate": {
						"type": "string"
					},
					"appointmentTime": {
						"type": "string"
					},
					"timeDuration": {
						"type": "string"
					},
					"appointmentLocation": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"purpose",
					"host",
					"department",
					"company",
					"appointmentDate",
					"appointmentTime",
					"timeDuration",
					"appointmentLocation"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateAppointmentAttributes": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"purpose": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"pending",
							"confirmed",
							"canceled",
							"onhold",
							"completed"
						]
					},
					"host": {
						"type": "string"
					},
					"department": {
						"type": "string"
					},
					"company": {
						"type": "string"
					},
					"appointmentDate": {
						"type": "string"
					},
					"appointmentTime": {
						"type": "string"
					},
					"timeDuration": {
						"type": "string"
					},
					"appointmentLocation": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "JWT token (Bearer <token>)"
			}
		}
	},
	"info": {
		"title": "Attendance Backend API",
		"version": "1.0.0",
		"description": "API documentation for attendance management system",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "API Support"
		}
	},
	"paths": {
...
		"/api/appointments": {
			"get": {
				"operationId": "GetAllAppointments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_AppointmentResponse-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Appointments"
				],
				"security": [
					{
						"jwt": [
							"appointment:read"
						]
					}
				],
				"parameters": []
			}
		},
		"/api/appointments/user/{userId}": {
			"get": {
				"operationId": "GetAppointmentsByUserId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_AppointmentResponse-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Appointments"
				],
				"security": [
					{
						"jwt": [
							"appointment:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/appointments/date/{date}": {
			"get": {
				"operationId": "GetAppointmentsByDate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_AppointmentResponse-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Appointments"
				],
				"security": [
					{
						"jwt": [
							"appointment:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "date",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/appointments/department/{department}": {
			"get": {
				"operationId": "GetAppointmentsByDepartment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_AppointmentResponse-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Appointments"
				],
				"security": [
					{
						"jwt": [
							"appointment:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "department",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/appointments/host/{host}": {
			"get": {
				"operationId": "GetAppointmentsByHost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_AppointmentResponse-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Appointments"
				],
				"security": [
					{
						"jwt": [
							"appointment:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "host",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/appointments/time/{time}": {
			"get": {
				"operationId": "GetAppointmentsByTime",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_AppointmentResponse-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Appointments"
				],
				"security": [
					{
						"jwt": [
							"appointment:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "time",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/appointments/location/{location}": {
			"get": {
				"operationId": "GetAppointmentsByLocation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_AppointmentResponse-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Appointments"
				],
				"security": [
					{
						"jwt": [
							"appointment:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "location",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/appointments/{id}": {
			"get": {
				"operationId": "GetAppointmentById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_AppointmentResponse-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"Appointments"
				],
				"security": [
					{
						"jwt": [
							"appointment:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteAppointment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Appointments"
				],
				"security": [
					{
						"jwt": [
							"appointment:delete"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/appointments/create": {
			"post": {
				"operationId": "CreateAppointment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_AppointmentResponse-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"Appointments"
				],
				"security": [
					{
						"jwt": [
							"appointment:create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AppointmentCreateAttributes"
							}
						}
					}
				}
			}
		},
		"/api/appointments/{id}/update": {
			"put": {
				"operationId": "UpdateAppointment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_AppointmentResponse-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"Appointments"
				],
				"security": [
					{
						"jwt": [
							"appointment:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateAppointmentAttributes"
							}
						}
					}
				}
			}
		},
		"/api/appointments/confirm/{id}": {
			"put": {
				"operationId": "ConfirmAppointment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_AppointmentResponse-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"Appointments"
				],
				"security": [
					{
						"jwt": [
							"appointment:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/appointments/cancel/{id}": {
			"put": {
				"operationId": "CancelAppointment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_AppointmentResponse-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"Appointments"
				],
				"security": [
					{
						"jwt": [
							"appointment:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/appointments/onhold/{id}": {
			"put": {
				"operationId": "OnholdAppointment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_AppointmentResponse-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"Appointments"
				],
				"security": [
					{
						"jwt": [
							"appointment:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/appointments/complete/{id}": {
			"put": {
				"operationId": "CompleteAppointment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_AppointmentResponse-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"Appointments"
				],
				"security": [
					{
						"jwt": [
							"appointment:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}